refactor(core): simplify event merge helper
